+++ src/pdflinkcheck/analyze_pypdf.py
@@ -10,7 +10,7 @@ from pypdf.generic import Destination, NameObject, ArrayObject, IndirectObject
 
 from pdflinkcheck.io import error_logger, export_report_data, get_first_pdf_in_cwd, LOG_FILE_PATH
 from pdflinkcheck.report import run_report
-from pdflinkcheck.validate import run_validation
+#from pdflinkcheck.validate import run_validation
 
 """
 Inspect target PDF for both URI links and for GoTo links, using only pypdf, not Fitz